feat(gradle): starting to extracting gradle scripts from build files
