+++ build.gradle.kts
@@ -4,8 +4,7 @@ import java.nio.charset.StandardCharsets
 
 @Suppress("DSL_SCOPE_VIOLATION")
 plugins {
-    alias(libs.plugins.kotlin)
-    id("com.diffplug.spotless") version "6.19.0"
+    `java`
 }
 
 val mavenGroup by properties
@@ -33,61 +32,13 @@ subprojects {
     apply(plugin = "kotlin")
     apply(plugin = "maven-publish")
     apply(plugin = "com.diffplug.spotless")
+    apply(plugin = "peripheralium.linting")
 
     this.extra["curseforgeKey"] = secretEnv["CURSEFORGE_KEY"] ?: System.getenv("CURSEFORGE_KEY") ?: ""
     this.extra["modrinthKey"] = secretEnv["MODRINTH_KEY"] ?: System.getenv("MODRINTH_KEY") ?: ""
     this.extra["rootProjectDir"] = rootProjectDir
 
     group = mavenGroup!!
-
-    val javaVersion = JavaVersion.VERSION_17
-    java {
-        toolchain { languageVersion.set(JavaLanguageVersion.of(javaVersion.toString())) }
-        sourceCompatibility = javaVersion
-        targetCompatibility = javaVersion
-        withSourcesJar()
-    }
-    tasks {
-        withType<JavaCompile> {
-            options.encoding = "UTF-8"
-            sourceCompatibility = javaVersion.toString()
-            targetCompatibility = javaVersion.toString()
-            options.release.set(javaVersion.toString().toInt())
-        }
-        withType<org.jetbrains.kotlin.gradle.tasks.KotlinCompile> {
-            kotlinOptions { jvmTarget = javaVersion.toString() }
-        }
-    }
-
-    spotless {
-        encoding = StandardCharsets.UTF_8
-        lineEndings = LineEnding.UNIX
-
-        fun FormatExtension.defaults() {
-            endWithNewline()
-            trimTrailingWhitespace()
-            indentWithSpaces(4)
-        }
-
-        java {
-            defaults()
-            removeUnusedImports()
-        }
-
-        val ktlintConfig = mapOf(
-            "ktlint_standard_no-wildcard-imports" to "disabled",
-        )
-
-        kotlinGradle {
-            defaults()
-            ktlint().editorConfigOverride(ktlintConfig)
-        }
-
-        kotlin {
-            defaults()
-            ktlint().editorConfigOverride(ktlintConfig)
-        }
-    }
 }
 
 repositories {
