+++ projects/core/build.gradle.kts
@@ -1,7 +1,7 @@
 @Suppress("DSL_SCOPE_VIOLATION")
 plugins {
     alias(libs.plugins.vanillaGradle)
-    alias(libs.plugins.kotlin)
+    id(libs.plugins.kotlin.get().pluginId) apply false
     idea
 }
